alterando métodos para melhor retorno dos métodos de contagem do dinheiro

+++ Money-Track/demo/src/main/java/com/money_track/demo/entities/DTO/LaunchesDataDTO.java
@@ -1,6 +1,7 @@
 package com.money_track.demo.entities.DTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record LaunchesDataDTO(List<LaunchDTO> launches, TypeValuesDTO typeValues, int totalLaunches, Double totalValue, List<CategoryTotalDTO> categoryTotals) {
+public record LaunchesDataDTO(List<LaunchDTO> launches, TypeValuesDTO typeValues, int totalLaunches, BigDecimal totalValue, List<CategoryTotalDTO> categoryTotals) {
 }
